models.pyをmodels配下のファイルに分割
