+++ app/models/booking.py
@@ -1,0 +1,16 @@
+from sqlalchemy import Column, Integer, ForeignKey, DateTime
+from sqlalchemy.orm import relationship
+from app.db.database import Base
+
+class Booking(Base):
+    __tablename__ = "bookings"
+
+    booking_id = Column(Integer, primary_key=True, index=True)
+    user_id = Column(Integer, ForeignKey("users.user_id", ondelete="SET NULL"), nullable=False)
+    room_id = Column(Integer, ForeignKey("rooms.room_id", ondelete="SET NULL"), nullable=False)
+    booked_num = Column(Integer)
+    start_datetime = Column(DateTime, nullable=False)
+    end_datetime = Column(DateTime, nullable=False)
+
+    user = relationship("User", back_populates="bookings")
+    room = relationship("Room", back_populates="bookings")
