+++ app/models/user.py
@@ -1,0 +1,11 @@
+from sqlalchemy import Column, Integer, String
+from sqlalchemy.orm import relationship
+from app.db.database import Base
+
+class User(Base):
+    __tablename__ = "users"
+
+    user_id = Column(Integer, primary_key=True, index=True)
+    username = Column(String(12), unique=True, index=True)
+
+    bookings = relationship("Booking", back_populates="user")
